refactor common code to shared.go
